+++ frameworks/native/hdiadapter/source/fast/fast_audio_capturer_source.cpp
@@ -398,8 +398,8 @@ int32_t FastAudioCapturerSourceInner::PrepareMmapBuffer()
         desc.transferFrameSize <= periodFrameMaxSize, ERR_OPERATION_FAILED,
         "ReqMmapBuffer invalid values: totalBufferFrames[%{public}d] transferFrameSize[%{public}d]",
         desc.totalBufferFrames, desc.transferFrameSize);
-    bufferTotalFrameSize_ = desc.totalBufferFrames; // 1440 ~ 3840
-    eachReadFrameSize_ = desc.transferFrameSize; // 240
+    bufferTotalFrameSize_ = static_cast<uint32_t>(desc.totalBufferFrames); // 1440 ~ 3840
+    eachReadFrameSize_ = static_cast<uint32_t>(desc.transferFrameSize); // 240
 
     CHECK_AND_RETURN_RET_LOG(frameSizeInByte <= ULLONG_MAX / bufferTotalFrameSize_,
         ERR_OPERATION_FAILED, "BufferSize will overflow!");
